ApiAi/requirements.txt
# requirements.txt
flask
pillow
ultralytics
numpy